fix : Trainer 회원 가입 양식 변경 및 layout , step 컴포넌트 분리작업 (#38)
* feat:Trainer 페이지 분리작업

* feat : signup 공통 파일 생성

* fix : trainer layout 변경

* fix : 폴더 구조 변경 및 불필요한 페이지 삭제

* fix : 불필요한 페이지 삭제

* fix : 피드백 변경사항 반영

* fix : async await 제거 처리
