+++ src/app/signup/type.d.ts
@@ -1,0 +1,5 @@
+import { PropsWithChildren } from "react";
+
+export type StepProps = {
+  next: () => void;
+} & PropsWithChildren;
