+++ src/utils/index.ts
@@ -1,0 +1,5 @@
+import { currentYear, months, years } from "./date";
+import { formatNumberWithComma, removeCommas } from "./format";
+
+export { currentYear, years, months };
+export { formatNumberWithComma, removeCommas };
